snowflake-export/pyproject.toml
[tool.poetry]
name = "snowflake-export"
version = "0.1"
description = "snowflake_export package to export table to snowflake"
authors = ["mopsde <v-ms@zillowgroup.com>"]


[[tool.poetry.source]]
name = "artifactory"
url = "https://artifactory.zgtools.net/artifactory/api/pypi/zg-pypi/simple"

[[tool.poetry.source]]
name = "public-pypi"
url = "https://pypi.org/simple"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
contract-registry-client-library =  { version = "2.28+master.x9d42118", source = "artifactory" }
urllib3 = {version = ">=1.0,<2.0", source = "public-pypi"}
requests = {version = "<2.30", source = "public-pypi"}
# ipython = "*"
pyspark = { version = ">=3.3.0", source = "public-pypi" }
snowflake-connector-python = {version = ">3.6.0", source = "public-pypi"}

[tool.poetry.scripts]
ingest_to_snowflake = "snowflake_export.ingest_to_snowflake:ingest_to_snowflake_entry"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
